backend/requirements.txt
async-timeout==5.0.1
bcrypt==4.0.1
bidict==0.23.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email-validator==2.2.0
eventlet==0.33.3
Flask==2.2.5
Flask-Cors==3.0.10
Flask-JWT-Extended==4.7.1
flask-mongoengine==1.0.0
Flask-SocketIO==5.3.4
flask-sqlalchemy==3.0.5
flask-wtf==1.2.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==3.0.2
mongoengine==0.29.1
PyJWT==2.10.1
pymongo==4.11.2
python-dotenv==1.0.0
python-engineio==4.11.2
python-socketio==5.12.1
redis==5.2.1
requests==2.31.0
simple-websocket==1.1.0
six==1.17.0
spotipy==2.23.0
sqlalchemy==2.0.39
typing-extensions==4.12.2
urllib3==2.3.0
werkzeug==3.1.3
wsproto==1.2.0
wtforms==3.2.1
zipp==3.21.0
